Improvements

+++ src/TinyTimer.cpp
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include "TinyTimer.h"
 
-void TinyTimer::on(const uint16_t match, TimerCallback callback) {
+void TinyTimer::on(const uint32_t match, TimerCallback callback) {
   onTimer = callback;
   setup(match);
 }
